pyproject.toml
[build-system]
requires = ["setuptools>64", "wheel"]
build_backend = "setuptools.build_meta"

[project]
name = "math_rag"
version = "0.1.0"
description = "GraphRAG for Mathematical Documents"
readme = "README.md"
authors = [
    { name = "Kai Jennissen", email = "k_jennissen@icloud.com" }
]
requires-python = ">=3.12, <3.13"
dependencies = [
    "coloredlogs",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "langchain-neo4j",
    "langgraph",
    "pandas",
    "python-dotenv",
    "pypdf",
    "pymupdf",
    "rich",
    "scikit-learn",
    "smolagents",
    "sqlmodel",
    "streamlit",
    "tabulate",
    "watchdog",
    "pyyaml",
    "detect-secrets>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "pytest",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.format]
quote-style = "double"
